Data Bidning logic was changed
Instead of binding to list of commands and binding using RelativeSource and finding Ancestors, i have come to a conclusion that it's much better to have some kind of Navigate Command field as part of ChildViewModel class itself. It adds more flexibility to application. Xaml structure was changed from ContentControl to ListBox. The main advantage of such implementation is that now i can track SelectedItem directly which makes my life easier.